pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-runner"
version = "0.1.0"
description = "LLM evaluation toolkit for running prompts across multiple models"
requires-python = ">=3.10"
dependencies = [
    "ollama",
    "pyyaml",
    "stable-diffusion-cpp-python",
]

[project.optional-dependencies]
dev = [
    "black",
    "mypy",
    "pillow",
    "pytest",
    "ruff",
    "types-PyYAML",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
prompt_runner = ["py.typed"]

[[tool.mypy.overrides]]
module = "stable_diffusion_cpp.*"
ignore_missing_imports = true